users-pages-data
This api works page wise i.e page1, 2, etc. So there should be a similar paging in our app.  Important to not that the number of pages are not fixed to 2, but its dynamic so paging should be dynamic as well
Use any ui library to display a list of user and having a paging option with it.